Split types in 2 files
